Support splitting output files
Added the option '-s' which directs the compiler to output multiple
files. This could potentially mean different things for each kind of
ouptut. It is currently only used for the verilog output mode, in which
case we set filenames based on the module name. In order for the '-s'
option to work, the '-o' argument shoudl be of the form
'<path>/*.<ext>'; files will be output to '<path>/<name>.<ext>'.