Fixes bug in verilog logic for handling self
The existing code strips all references to `self` in names, however this
is bad in the case that `self` is actualy part of the name (e.g. a port
named `self_I`).  This updates the logic to only remove `self` when it
is the first element of a select path, rather than search/replacing the
entire select string.