Preserve port order in module instantiation
Updates for https://github.com/leonardt/verilogAST-cpp/pull/39 so that
the order of ports for module instances are preserved.  The old logic
used a std::map which uses lexical order, causing the output to sort
the ports based on the names.  Migrating to a std::vector preserve's
CoreIR's internal ordering of the ports (which is based on insertion
order)