Refactor libraries into filegroups
Due to circular dependencies in the sub-libraries (e.g. between ir/ and
passes/), *and* since bazel does not currently allow -Wl,--start-group,
we must build libcoreir.so monolithically. That is, we can not build
libir.a and libpasses.a and combine them into libcoreir.so. This is
actually the same as the old cmake build system, so it is not a
regression.