Adds support for magma's user defined namespace (#922)
Magma added support for a user_namespace, but currently magma only
compiles the namespace containing the top module.  This doesn't work
when we want some modules in the user_namespace, but other modules in
the global namespace (e.g. imported verilog modules that shouldn't have
the user_namespace prefix).  These changes are required so that magma
can compile the whole context instead of just the namespace of the top
module.

* Add Context.setTop to C API (so magma can set the top module in a
  context)
* Add saveToFile of context to C API (so magma can save an entire
  context instead of just the namespace of the top module)
* Add option to saveToFile to skip default imported library namespaces
  like mantle and commonlib.  These don't need to be in the JSON (and
  break downstream gold files when switching to this flow), skipping
  them improves readability (less noise).
* Adds logic in json pass to skip a namespace if it's empty (e.g. global
  namespace may be empty when magma is using only the user_namespace)