Avoid name conflicts between instance port wires and instance names.
Fixes https://github.com/rdaly525/coreir/issues/954.

When declaring connections, we now check the instance names to see if
there's a conflict, if there is, we generate a unique name.  We then
insert an entry into a map from original instance port wire names to
uniquified names.  When code generating wire references, we lookup in
this map to see if a name has been uniquified.

While this generally works, the logic inside the selectpath -> verilog
reference is a bit complicated.  Basically, we build up an identifier
incrementally (instance name, then port name, or in another case just
instance name for wires).  Sometimes these are used inside index, slice,
or attribute nodes, so we have to check all these cases and only lookup
the mapping at the appropriate time (e.g. when we're creating an index,
we know the identifier that is being indexed should be mapped, but we
don't need to map nested index nodes).  This adds a bunch of cases
inside this conversion logic making the code complex.  I will see if I
can refactor the logic to tame the complexity, I think perhaps one
option is to build a "base" reference (identifier out of instance name +
port name, or instance name) and map that, then add the extra wrappers
(index, slice, attributes).  The one challenge is the ordering of the
wrappers being inconsistent (index/slices happen at the end of the
select path, attributes from hierarchical selects happen before the
final instance/port reference) so I'm not sure if we'll be able to avoid
the two different cases there.