Add SyncReadMemory primitive
This adds a SyncReadMemory primitive that is similar to coreir.mem with
the sync_read parameter enabled, except that it also supports a read
enable port.

The motivation is that we want to generate the `ren` logic in the
verilog using the `if` inside the `always` block for synthesis.
However, we can't smash this inside the existing memory generator since
it would require an "optional" `ren` port that isn't used when sync_read
is disabled.  It doesn't seem like verilog supports this optional port
pattern.

One alternative to defining a separate primitive would be to always
generate a `ren` port in the verilog and ignore it if `sync_read` is
disabled.

Ideally we'd have this in place of the sync_read parameter, but I left
the old logic in for backwards compatability.  With sync_read, if the
user doesn't want to use the enable logic, we can just set the input to
1 and presumably the downstream tool can optimize that out.

Setting ren to a constant 1 when the enable logic is not desired seemed
like a better pattern than having a dangling `ren` port when there's a
combinational read, but the downside is we have to use a separate
primitive.

Perhaps an alternative long-term solution would be to add a
`sync_read_enable` argument to the memory and have it generate different
verilog code based on the parameter.  This might be a good effort for
other cases, e.g. where we use `generate if` for the verilog memory init
logic (instead we could just generate the `init` code if the parameter
is enabled), but this requires quite a bit of work since the verilog
primitive code generation wasn't designed to support this (right now we
just store a static verilog string to emit, rather than generating the
string dynamically, we'd need to create some API for this).