Add suport for .sv extension (#978)
* Add suport for .sv extension

Needed for https://github.com/phanrahan/magma/issues/880

Allows the user to choose `.sv` as an output.  In the future, we should
consider forcing the `disable-ndarray` option when `.v` is used (i.e.
choosing `.sv` enables this feature since it should be compatible with
tools that support `.sv`).   However, for backwards compatability, I
think we should leave the behavior as is (or else various downstream
users will have different code generated by default since everyone is
using `.v` right now but some are expected the ndarray feature enabled
for `.v`).

* Update test code for API change