Add support for verilog body string (#987)
This was a feature we used to support in magma (see first example
https://github.com/phanrahan/magmathon/blob/master/notebooks/advanced/verilog.ipynb)
that used the verilog backend.  This adds support for the same feature
in the coreir backend.

This simply allows the user to define circuit interface in magma/coreir,
but the body in verilog.  This is slightly different than inline verilog
in that we allow the user to drive signals from the verilog body string.
Otherwise, if we drove a signal from inline_verilog, we'd get an
undriven error in magma/coreir since we aren't aware of the driver from
the string.