Support multi-level inlining
This commit adds support to the SymbolTableInterface API to allow for
multi-level inlining. It does this by making the return type of
get_inlined_instance_name() to be Union[str, Sentinel] (just like
get_instance_name()). We also update the internal data structures to
support storing this union, as well as ser/deser.

NOTE: We still have not instrumented the inline pass.