Instrument inline symbol table (#997)
* Add back flattentypes test

In a force push, commit 0e1b09ea3e00638cdd4a57a4db5c41641811b288 was
lost. This commit adds it back.

Original commit message:

"adds basic symbol table test for flatten types"

* Add simple inlineInstance test

* Move set-up and tear-down to test fixture

* Add inlineInstance symbol table test

NOTE: This is just the test for inlining. It does not yet instrument
the pass to update the symbol table.

* Support multi-level inlining

This commit adds support to the SymbolTableInterface API to allow for
multi-level inlining. It does this by making the return type of
get_inlined_instance_name() to be Union[str, Sentinel] (just like
get_instance_name()). We also update the internal data structures to
support storing this union, as well as ser/deser.

NOTE: We still have not instrumented the inline pass.

* Add instance type API

Add API to allow for get/setting types of instances (a simple str ->
str -> str function).

* Update inlined instance API to take a vector of instances

* Fix typo

setInlineInstance -> setInlinedInstance

* Update inline test for new API

* Update simple hierarchy example

* Update inline test for new hierarchy example

* Revert inlinedInstance API change

This change reverts the change to make the {set,get}inlinedInstance API
take in a vector. We now take in a pair as before.

* Add scaffolding for logging

This change introduces logging APIs to the symbol table to allow for
staged symbol table updates.

TODO: Implement the logic for finalize() (merging logs) and pass the
right arguments into the logging call from inlineInstance().

* Fix basic inline test

* Make logger accesible directly from symbol table

* Update logger interface

* Log new instances in module def

* Log inlined instances in inlineInstance()

* Add debugging capabilities to logger

* Skip logging new/removed passthrough instances

* Fix debug iterator bug

* Pause logging during extraneous actions in inlineInstances

* Implement symbol table logging finalization

* Set types during finalization

* Fix compiler issues/warnings

* Fix compiler issues/warnings

* Put symbol table behind flag

* Fix minor issues